+++ aderyn_core/src/ast/impls/node/enumerations.rs
@@ -1,17 +1,7 @@
-use super::*;
+use crate::ast::*;
 use crate::visitor::ast_visitor::*;
 use eyre::Result;
-use serde::{Deserialize, Serialize};
 use std::fmt::Display;
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct EnumValue {
-    pub name: String,
-    pub name_location: Option<String>,
-    pub src: String,
-    pub id: NodeID,
-}
 
 impl Node for EnumValue {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
@@ -28,17 +18,6 @@ impl Display for EnumValue {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_str(self.name.as_str())
     }
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct EnumDefinition {
-    pub name: String,
-    pub name_location: Option<String>,
-    pub members: Vec<EnumValue>,
-    pub canonical_name: Option<String>,
-    pub src: String,
-    pub id: NodeID,
 }
 
 impl Node for EnumDefinition {
