+++ aderyn_core/src/ast/impls/node/mod.rs
@@ -1,3 +1,4 @@
 pub mod blocks;
 pub mod contracts;
 pub mod documentation;
+pub mod enumerations;
